test: Add jasmine unit test framework to the tet serializer
