+++ package.json
@@ -16,7 +16,8 @@
   "devDependencies": {
     "@types/electron-devtools-installer": "^2.0.2",
     "@types/electron-settings": "^3.0.0",
-    "@types/node": "8.0.28",
+    "@types/jasmine": "2.6.0",
+    "@types/node": "8.0.30",
     "@types/node-hid": "^0.5.2",
     "@types/usb": "^1.1.3",
     "autoprefixer": "^6.5.3",
@@ -63,7 +64,6 @@
     "build:electron": "cross-env AOT_BUILD=true run-s -sn build:electron:renderer build:electron:main",
     "build:electron:main": "lerna exec --scope uhk-agent npm run build",
     "build:electron:renderer": "lerna exec --scope uhk-web npm run build:renderer",
-    "build:test": "lerna exec --scope test-serializer npm run build",
     "build:common": "lerna exec --scope uhk-common npm run build",
     "build:usb": "lerna exec --scope uhk-usb npm run build",
     "server:web": "lerna exec --scope uhk-web npm start",
