+++ packages/test-serializer/package.json
@@ -16,10 +16,16 @@
   "dependencies": {
   },
   "devDependencies": {
+    "@types/jasmine": "^2.6.0",
+    "@types/node": "8.0.30",
+    "jasmine": "^2.8.0",
+    "jasmine-core": "^2.8.0",
+    "jasmine-node": "2.0.0",
+    "jasmine-ts": "^0.2.1",
+    "ts-node": "3.3.0",
     "uhk-common": "1.0.0"
   },
   "scripts": {
-    "build": "webpack",
-    "test": "node ./test-serializer.js"
+    "test": "jasmine-ts --config=jasmine.json"
   }
 }
